add indent message

+++ app/controllers/IndentController.php
@@ -19,7 +19,17 @@ class IndentController extends BaseController {
 		$indent->user_id = $user->id;
  		
 		if($indent->save())
+		{
+			$content = '用户' . $user->name . '对你的' . $product->name . '商品下了订单!';
+			$message = Message::create(array(
+					'title' => '下单消息',
+					'content' => $content,
+					'sender' => 8,
+					'receiver' => $product->shop->user_id,
+					'type' => 1
+				));
 			return Response::json(array('errCode' => 0,'message' => '创建订单成功!', 'newIndentId' => $indent->id));
+		}			
 
 		return Response::json(array('errCode' => 3,'message' => '创建订单失败!'));
 	}
@@ -29,14 +39,28 @@ class IndentController extends BaseController {
 		$indentId = Input::get("indentId");
 		$user = Sentry::getUser();
 
-		$indent = Indent::find($indentId);
+		$indent = Indent::with('product.shop')->find($indentId);
 		if(!isset($indent))
 			return Response::json(array('errCode' => 1,'message' => '该订单不存在!'));
 
-		if($indent->delete())
-			return Response::json(array('errCode' => 0,'message' => '取消订单成功!'));
+		if($indent->user_id != $user->id)
+			return Response::json(array('errCode' => 2,'message' => '你没有此操作权限!'));
 
-		return Response::json(array('errCode' => 2,'message' => '取消订单失败!'));
+		$content = '用户' . $user->name . '取消了你的' .  $indent->product->name . '商品的订单!';
+		$receiver = $indent->product->shop->user_id;
+		if($indent->delete())
+		{		
+			$message = Message::create(array(
+					'title' => '取消订单消息',
+					'content' => $content,
+					'sender' => 8,
+					'receiver' => $receiver,
+					'type' => 1
+				));
+			return Response::json(array('errCode' => 0,'message' => '取消订单成功!'));
+		}	
+
+		return Response::json(array('errCode' => 3,'message' => '取消订单失败!'));
 	}
 
 	public function getIndent()
